projects/hcp/packages/rahcp-cli/pyproject.toml
[project]
name = "rahcp-cli"
version = "0.1.0"
description = "CLI for HCP Unified API"
requires-python = ">=3.13"
dependencies = ["rahcp-client", "rahcp-tracker", "rahcp-iiif", "typer>=0.15", "rich>=13.0", "pyyaml>=6.0"]

[project.optional-dependencies]
validate = ["rahcp-validate"]

[project.scripts]
rahcp = "rahcp_cli.main:app"

[tool.uv.sources]
rahcp-client = { workspace = true }
rahcp-iiif = { workspace = true }
rahcp-tracker = { workspace = true }
rahcp-validate = { workspace = true }

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
